chore: abstract validation
