+++ package.json
@@ -143,6 +143,7 @@
     "babel-preset-jest": "24.6.0",
     "codecov": "3.5.0",
     "commitizen": "4.0.3",
+    "commitlint": "^8.1.0",
     "commitlint-config-cz": "0.12.0",
     "conventional-changelog-cli": "2.0.21",
     "cz-customizable": "6.2.0",
